kafka_consumer/requirements.txt
kafka-python==2.0.2
requests==2.32.3 # Untuk HTTP request langsung ke Supabase
msgpack==1.0.8
